CoinFLip

+++ test/SimpleWallet/SimpleWalletHuff.t.sol
@@ -15,20 +15,18 @@ contract SimpleWalletHuffTest is Test {
         address addr = HuffDeployer.config().with_args(abi.encode(address(this))).deploy("SimpleWallet/SimpleWallet");
         simpleWallet = ISimpleWallet(addr);
         receiverWallet = new Receiver();
-
     }
 
     function testSimpleWallMakeTxHuff() public {
         uint256 _val = 1 ether;
         hoax(address(this));
-        uint x = simpleWallet.makeTx{value: _val}(address(receiverWallet), _val, abi.encode(_selector));
+        uint256 x = simpleWallet.makeTx{value: _val}(address(receiverWallet), _val, abi.encode(_selector));
         assertEq(_val, address(receiverWallet).balance);
         assertEq(x, _val);
     }
-    
+
     function testSimpleWallMakeTxHuffOwn() public {
         // emit log_bytes32(vm.load(address(simpleWallet), bytes32(0)));
         simpleWallet.owner();
     }
-
 }
